renamed test fixtures file
