+++ tests/test_wikitext_to_structured_sections.py
@@ -4,7 +4,7 @@ Inspiration for test cases drawn from,
  * https://en.wikipedia.org/wiki/Help:Wikitext#Links_and_URLs
 """
 from mwtext import Wikitext2StructuredSections
-from .mwpfh_wikilink_fixtures import WIKILINK_TEST_FIXTURES
+from .wikilink_fixtures import WIKILINK_TEST_FIXTURES
 
 
 def test_wikilinks_only():
